Adjusted cmsAdapter.js
